+++ src/anticarium_desktop/Jttp.cpp
@@ -68,6 +68,10 @@ void JTTP::onDataArrived(QNetworkReply* reply) {
     }
 }
 
+void JTTP::onSendData(const shared_types::Regime& regime) {
+    httpSend(REQUEST_TYPE::SEND, REQUEST_DATA::REGIME, regime);
+}
+
 void JTTP::onSendData(const shared_types::Control& control) {
     httpSend(REQUEST_TYPE::SEND, REQUEST_DATA::CONTROL_DATA, control);
 }
@@ -94,10 +98,12 @@ void JTTP::httpSend(REQUEST_TYPE requestType, REQUEST_DATA requestData, const nl
         networkAccessManager->get(networkRequest);
     }
 
-    // Instantly process requests so they synchronously
-    QCoreApplication::processEvents(QEventLoop::ExcludeUserInputEvents);
+    // Wait for request to finish executing
+    QEventLoop loop;
+    connect(networkAccessManager, &QNetworkAccessManager::finished, &loop, &QEventLoop::quit);
+    loop.exec();
 
-    SPDLOG_INFO(QString("Data requested from url %1").arg(url).toStdString());
+    SPDLOG_INFO(QString("Data %1, url: %2").arg(requestTypeString, url).toStdString());
 }
 
 JTTP::~JTTP() {